requirements.txt
# 核心依赖
httpx>=0.27.0           # 异步HTTP客户端
feedparser>=6.0.11      # RSS解析
beautifulsoup4>=4.12.0  # HTML解析
lxml>=5.1.0             # 高性能XML/HTML解析器

# 工具库
python-dateutil>=2.8.0  # 日期解析
loguru>=0.7.0           # 日志
schedule>=1.2.0         # 定时调度
pytz>=2024.1            # 时区处理

# Web 展示（可选）
fastapi>=0.115.0        # Web框架
uvicorn>=0.32.0         # ASGI服务器

# 可选：中文NLP
jieba>=0.42.1           # 中文分词（用于关键词过滤）
